Intengo de implementacion de seeds - no funciona

+++ source/database/migrations/20220820004814-cart.js
@@ -5,7 +5,7 @@ module.exports = {
     try {
       await queryInterface.createTable("cart", {
         id: {
-          type: Sequelize.SMALLINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
           autoIncrement: true,
           primaryKey: true,
